Moves sources to their own folder
